requirements.txt
streamlit==1.31.0
pandas==2.1.0
numpy==1.26.0
matplotlib==3.8.0
Pillow==10.0.0
streamlit-extras==0.3.5
streamlit-lottie==0.0.5
requests==2.31.0
groq==0.4.0
python-dotenv==1.0.0
